Extract import handling into modifier
Introduce MenuBarImportHandlerModifier (Content+ImportFiles.swift) to handle .shouldHandleImport notifications and import files via FileState, binding the window and showing alerts on error. Apply the modifier in ContentView and remove the previous onReceive handler and handleImport implementation from ContentView to centralize import logic. The modifier includes platform-specific window binding (AppKit/UIKit) and performs imports on a detached Task.
